# api/openapi/api-gateway.yaml
openapi: 3.1.0
info:
  title: Personal Finance OS API Gateway
  version: 0.4.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/analytics/projections:
    get:
      summary: Inspect analytics projection metadata
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Analytics projection metadata
  /api/v1/analytics/projections/daily-spend:
    get:
      summary: Query daily spend projections
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: category
          schema:
            type: string
      responses:
        '200':
          description: Daily spend projection
  /api/v1/analytics/projections/alerts:
    get:
      summary: Query alert projections
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
        - in: query
          name: severity
          schema:
            type: string
      responses:
        '200':
          description: Alert projection
  /api/v1/analytics/projections/summary:
    get:
      summary: Query summary analytics projection
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Summary projection
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...